+++ src/main/kotlin/dev/d1s/linda/dto/converter/impl/redirect/RedirectDtoConverter.kt
@@ -17,7 +17,7 @@
 package dev.d1s.linda.dto.converter.impl.redirect
 
 import dev.d1s.linda.dto.redirect.RedirectDto
-import dev.d1s.linda.entity.Redirect
+import dev.d1s.linda.entity.redirect.Redirect
 import dev.d1s.teabag.dto.DtoConverter
 import dev.d1s.teabag.stdlib.collection.mapToSet
 import org.springframework.stereotype.Component
