+++ src/main/kotlin/dev/d1s/linda/dto/converter/impl/redirect/RedirectUpdateDtoConverter.kt
@@ -17,7 +17,7 @@
 package dev.d1s.linda.dto.converter.impl.redirect
 
 import dev.d1s.linda.dto.redirect.RedirectUpdateDto
-import dev.d1s.linda.entity.Redirect
+import dev.d1s.linda.entity.redirect.Redirect
 import dev.d1s.linda.service.ShortLinkService
 import dev.d1s.linda.strategy.shortLink.byId
 import dev.d1s.teabag.dto.DtoConverter
