+++ src/main/kotlin/dev/d1s/linda/entity/utmParameter/UtmParameter.kt
@@ -16,7 +16,7 @@
 
 package dev.d1s.linda.entity.utmParameter
 
-import dev.d1s.linda.entity.Redirect
+import dev.d1s.linda.entity.redirect.Redirect
 import dev.d1s.linda.entity.ShortLink
 import dev.d1s.linda.entity.common.Identifiable
 import javax.persistence.Column
