+++ src/main/kotlin/dev/d1s/linda/repository/RedirectRepository.kt
@@ -16,7 +16,7 @@
 
 package dev.d1s.linda.repository
 
-import dev.d1s.linda.entity.Redirect
+import dev.d1s.linda.entity.redirect.Redirect
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
