+++ src/main/kotlin/dev/d1s/linda/service/RedirectService.kt
@@ -17,7 +17,7 @@
 package dev.d1s.linda.service
 
 import dev.d1s.linda.dto.redirect.RedirectDto
-import dev.d1s.linda.entity.Redirect
+import dev.d1s.linda.entity.redirect.Redirect
 import dev.d1s.linda.entity.utmParameter.UtmParameter
 import dev.d1s.teabag.dto.EntityWithDto
 import dev.d1s.teabag.dto.EntityWithDtoSet
