+++ src/main/kotlin/dev/d1s/linda/service/impl/BaseInterfaceServiceImpl.kt
@@ -19,7 +19,7 @@ package dev.d1s.linda.service.impl
 import dev.d1s.linda.configuration.properties.BaseInterfaceConfigurationProperties
 import dev.d1s.linda.configuration.properties.SslConfigurationProperties
 import dev.d1s.linda.constant.mapping.BASE_INTERFACE_CONFIRMATION_SEGMENT
-import dev.d1s.linda.entity.Redirect
+import dev.d1s.linda.entity.redirect.Redirect
 import dev.d1s.linda.entity.utmParameter.UtmParameterType
 import dev.d1s.linda.service.*
 import dev.d1s.teabag.web.buildFromCurrentRequest
